// src/app/shared/interfaces/views.enum.ts
export enum viewsEnum {
  COMPANIES = 'COMPANIES',
  USERS = 'USERS',
  TARIFFS = 'TARIFFS',
  STATISTICS = 'STATISTICS',

  COMPANY_USERS = 'COMPANY_USERS',
  COMPANY_USER_TARIFFS = 'COMPANY_USER_TARIFFS',
  COMPANY_USER_STATISTICS = 'COMPANY_USER_STATISTICS',
  COMPANY_CLIENTS = 'COMPANY_CLIENTS'
}